Move activity metadata testing
