+++ sqlserver/tests/test_statements.py
@@ -1,7 +1,6 @@
 # -*- coding: utf-8 -*-
 from __future__ import unicode_literals
 
-import concurrent
 import logging
 import math
 import os
@@ -21,7 +20,6 @@ from datadog_checks.sqlserver import SQLServer
 from datadog_checks.sqlserver.statements import SQL_SERVER_QUERY_METRICS_COLUMNS, obfuscate_xml_plan
 
 from .common import CHECK_NAME
-from .conftest import DEFAULT_TIMEOUT
 
 try:
     import pyodbc
@@ -279,8 +277,8 @@ def test_statement_metrics_and_plans(
     "metadata,expected_metadata_payload",
     [
         (
-            {'tables_csv': 'ϑings', 'commands': ['SELECT'], 'comments': ['-- Test comment']},
-            {'tables': ['ϑings'], 'commands': ['SELECT'], 'comments': ['-- Test comment']},
+            {'tables_csv': 'sys.databases', 'commands': ['SELECT'], 'comments': ['-- Test comment']},
+            {'tables': ['sys.databases'], 'commands': ['SELECT'], 'comments': ['-- Test comment']},
         ),
         (
             {'tables_csv': '', 'commands': None, 'comments': None},
@@ -288,31 +286,18 @@ def test_statement_metrics_and_plans(
         ),
     ],
 )
-def test_statement_metadata(aggregator, dd_run_check, dbm_instance, datadog_agent, metadata, expected_metadata_payload):
-    # Enable activity for this test to check for metadata
-    dbm_instance['query_activity'] = {'enabled': True, 'run_sync': True, 'collection_interval': 0.1}
-
-    def _get_conn_for_user(user):
-        conn_str = 'DRIVER={};Server={};Database=master;UID={};PWD={};'.format(
-            dbm_instance['driver'], dbm_instance['host'], user, "Password12!"
-        )
-        conn = pyodbc.connect(conn_str, timeout=DEFAULT_TIMEOUT, autocommit=False)
-        conn.timeout = DEFAULT_TIMEOUT
-        return conn
-
+def test_statement_metadata(
+    aggregator, dd_run_check, dbm_instance, bob_conn, datadog_agent, metadata, expected_metadata_payload
+):
     check = SQLServer(CHECK_NAME, {}, [dbm_instance])
 
     query = '''
     -- Test comment
-    SELECT * FROM ϑings'''
-    query_signature = '5ff34f4267b108c6'
+    select * from sys.databases'''
+    query_signature = 'ee1663c796378ab0'
 
-    bob_conn = _get_conn_for_user("bob")
-
-    def _run_test_query(conn, q):
-        cur = conn.cursor()
-        cur.execute("USE {}".format("datadog_test"))
-        cur.execute(q)
+    def _run_query():
+        bob_conn.execute_with_retries(query)
 
     def _obfuscate_sql(sql_query, options=None):
         return json.dumps({'query': sql_query, 'metadata': metadata})
@@ -320,10 +305,9 @@ def test_statement_metadata(aggregator, dd_run_check, dbm_instance, datadog_agen
     # Execute the query with the mocked obfuscate_sql. The result should produce an event payload with the metadata.
     with mock.patch.object(datadog_agent, 'obfuscate_sql', passthrough=True) as mock_agent:
         mock_agent.side_effect = _obfuscate_sql
-        _run_test_query(bob_conn, query)
+        _run_query()
         dd_run_check(check)
-        # Second run will emit metrics based on the diff of the current and prev state
-        _run_test_query(bob_conn, query)
+        _run_query()
         dd_run_check(check)
 
     dbm_samples = aggregator.get_event_platform_events("dbm-samples")
@@ -352,43 +336,6 @@ def test_statement_metadata(aggregator, dd_run_check, dbm_instance, datadog_agen
     metric = matching_metrics[0]
     assert metric['dd_tables'] == expected_metadata_payload['tables']
     assert metric['dd_commands'] == expected_metadata_payload['commands']
-
-    # Setup activity for metadata testing
-    blocking_query = "INSERT INTO ϑings WITH (TABLOCK, HOLDLOCK) (name) VALUES ('puppy')"
-    fred_conn = _get_conn_for_user('fred')
-
-    def _run_query_with_mock_obfuscator(conn, query):
-        # Execute the query with the mocked obfuscate_sql. The result should produce an event payload with the metadata.
-        with mock.patch.object(datadog_agent, 'obfuscate_sql', passthrough=True) as mock_agent:
-            mock_agent.side_effect = _obfuscate_sql
-            _run_test_query(conn, query)
-
-    # bob's query blocks until the tx is completed, so it needs to be run asynchronously.
-    executor = concurrent.futures.ThreadPoolExecutor(1)
-    executor.submit(_run_query_with_mock_obfuscator, bob_conn, blocking_query)
-
-    # fred's query will get blocked by bob, so it needs to be run asynchronously.
-    executor.submit(_run_query_with_mock_obfuscator, fred_conn, query)
-
-    dd_run_check(check)
-
-    bob_conn.commit()
-    bob_conn.close()
-    fred_conn.close()
-    executor.shutdown(wait=True)
-
-    dbm_activity = aggregator.get_event_platform_events("dbm-activity")
-    assert dbm_activity, "should have collected at least one activity"
-    matching_activity = []
-    for event in dbm_activity:
-        for activity in event['sqlserver_activity']:
-            if activity['query_signature'] == query_signature:
-                matching_activity.append(activity)
-    assert len(matching_activity) == 1
-    activity = matching_activity[0]
-    assert activity['dd_tables'] == expected_metadata_payload['tables']
-    assert activity['dd_commands'] == expected_metadata_payload['commands']
-    assert activity['dd_comments'] == expected_metadata_payload['comments']
 
 
 @pytest.mark.integration
